+++ src/app/models/action-type.ts
@@ -1,0 +1,6 @@
+export enum ActionType {
+    Cancel,
+    Delete,
+    Success,
+    Update,
+}