+++ src/app/tools/gen-mapper/node-drawer/node-drawer.component.ts
@@ -3,17 +3,21 @@ import { FormControl, FormGroup } from '@angular/forms';
 import { MatDialog } from '@angular/material/dialog';
 import { MatDrawer } from '@angular/material/sidenav';
 import { LocaleService } from '@core/locale.service';
+import { OtherPeopleGroup, UnknownPeopleGroup } from '@core/people-group.service';
 import { Unsubscribable } from '@core/Unsubscribable';
+import { ActionType } from '@models/action-type';
 import { DocumentDto } from '@models/document.model';
-import { NodeDto } from '@models/node.model';
+import { NodeDto, PeopleAttributes } from '@models/node.model';
+import { Template } from '@models/template.model';
 import { FileInputDialogComponent } from '@shared/file-input-dialog/file-input-dialog.component';
-import { GMField } from '@templates';
-import { assign } from 'lodash';
-import { takeUntil } from 'rxjs/operators';
+import { ControlType, GMField } from '@templates';
+import { assign, cloneDeep } from 'lodash';
+import { filter, takeUntil } from 'rxjs/operators';
 import { ConfirmDialogComponent } from '../dialogs/confirm-dialog/confirm-dialog.component';
+import { PeopleDialogComponent, PeopleDialogConfig, PeopleDialogResponse } from '../dialogs/people-dialog/people-dialog.component';
+import { AddPeopleGroupConfig, SelectPeopleGroupDialogComponent } from '../dialogs/select-people-group-dialog/select-people-group-dialog.component';
 import { GenMapperService } from '../gen-mapper.service';
 import { NodeClipboardService } from '../node-clipboard.service';
-
 
 @Component({
     selector: 'app-node-drawer',
@@ -25,6 +29,12 @@ export class NodeDrawerComponent extends Unsubscribable implements OnInit {
     public nodes: NodeDto[];
     public document: DocumentDto;
     public documents: DocumentDto[];
+    public template: Template;
+    public isNodeInClipboard: boolean;
+    public fields: GMField[];
+    public form: FormGroup;
+    public pendingPeoples: PeopleAttributes[] = [];
+    public selectedTabIndex = 0;
 
     @Input()
     public hideActions: boolean;
@@ -47,11 +57,6 @@ export class NodeDrawerComponent extends Unsubscribable implements OnInit {
     @Output()
     public importSubtree = new EventEmitter<NodeDto>();
 
-    public isNodeInClipboard: boolean;
-    public clonedNode: NodeDto;
-    public fields: GMField[];
-    public form: FormGroup;
-
     constructor(
         private localeService: LocaleService,
         private genMapper: GenMapperService,
@@ -66,6 +71,7 @@ export class NodeDrawerComponent extends Unsubscribable implements OnInit {
             .pipe(takeUntil(this.unsubscribe))
             .subscribe(template => {
                 this.fields = template.fields;
+                this.template = template;
                 this.initializeForm();
             });
 
@@ -109,6 +115,7 @@ export class NodeDrawerComponent extends Unsubscribable implements OnInit {
 
     public configureNode(): void {
         this.checkClipboard();
+        this.pendingPeoples = cloneDeep(this.node.attributes.peoples);
         this.form.reset(this.node.attributes);
         this.form.patchValue({ parentId: this.node.parentId });
     }
@@ -148,9 +155,20 @@ export class NodeDrawerComponent extends Unsubscribable implements OnInit {
         }
 
         assign(this.node.attributes, value);
-
+        this.node.attributes.peoples = this.pendingPeoples;
         this.node.parentId = value.parentId;
         delete this.node.attributes.parentId;
+
+        // Apply Peoples Attributes to field Attributes
+        const peopleField = this.template.getField('peoples');
+        peopleField.fields.forEach(field => {
+            this.node.attributes[field.id] = 0;
+        });
+        this.node.attributes.peoples.forEach(people => {
+            peopleField.fields.forEach(field => {
+                this.node.attributes[field.id] += people[field.id];
+            });
+        });
 
         this.updateNode.emit(this.node);
         this.drawer.disableClose = false;
@@ -227,6 +245,89 @@ export class NodeDrawerComponent extends Unsubscribable implements OnInit {
         this.drawer.disableClose = this.form.dirty;
     }
 
+    public selectPeople(people: PeopleAttributes): void {
+        this.dialog
+            .open<PeopleDialogComponent, PeopleDialogConfig, PeopleDialogResponse>(
+                PeopleDialogComponent, {
+                autoFocus: false,
+                data: {
+                    template: this.template,
+                    people: people
+                }
+            })
+            .afterClosed()
+            .pipe(filter(r => !!r))
+            .subscribe(result => {
+                if (result.actionType === ActionType.Success) {
+                    Object.assign(people, result.people);
+                    this.pendingPeoples = this.pendingPeoples.slice();
+                    this.form.markAsDirty();
+                }
+
+                if (result.actionType === ActionType.Update) {
+                    this.changePeopleGroup(people);
+                }
+
+                if (result.actionType === ActionType.Delete) {
+                    this.removePendingPeoples(result.people);
+                }
+            });
+    }
+
+    public addPeopleGroup(): void {
+        const countryField = this.fields.find(field => field.type === ControlType.countrySelector);
+
+        if (!countryField) {
+            return;
+        }
+
+        this.dialog
+            .open<SelectPeopleGroupDialogComponent, AddPeopleGroupConfig, PeopleAttributes>(
+                SelectPeopleGroupDialogComponent, {
+                data: {
+                    countryCode: this.form.get(countryField.id).value,
+                    template: this.template,
+                    peoples: this.pendingPeoples
+                }
+            })
+            .afterClosed()
+            .subscribe(result => {
+                if (result) {
+                    this.insertPendingPeoples(result);
+                    this.selectPeople(result);
+                }
+            })
+    }
+
+    public changePeopleGroup(people: PeopleAttributes): void {
+        const countryField = this.fields.find(field => field.type === ControlType.countrySelector);
+
+        if (!countryField) {
+            return;
+        }
+
+        const country = people.placeOfOrigin || this.form.get(countryField.id).value;
+
+        this.dialog
+            .open<SelectPeopleGroupDialogComponent, AddPeopleGroupConfig, PeopleAttributes>(
+                SelectPeopleGroupDialogComponent, {
+                data: {
+                    countryCode: country,
+                    template: this.template,
+                    peoples: this.pendingPeoples,
+                    selectedPeople: people,
+                }
+            })
+            .afterClosed()
+            .subscribe(result => {
+                if (result) {
+                    this.removePendingPeoples(people);
+                    this.insertPendingPeoples(result);
+                    this.insertUnknownPeopleGroupAtBeginning();
+                }
+            })
+    }
+
     private initializeForm(): void {
         const group: any = {};
         const fields: { name: string, order: number }[] = [];
@@ -275,5 +376,52 @@ export class NodeDrawerComponent extends Unsubscribable implements OnInit {
                 this.isNodeInClipboard = true;
             }
         }
+    }
+
+    private insertPendingPeoples(people: PeopleAttributes): void {
+        let found;
+
+        if (people.identifier === OtherPeopleGroup.PEID) {
+            found = this.pendingPeoples.find(f => f.identifier === people.identifier && f.placeOfOrigin === people.placeOfOrigin && f.label === people.label);
+        } else {
+            found = this.pendingPeoples.find(f => f.identifier === people.identifier);
+        }
+
+        if (found) {
+            Object.assign(found, people);
+        } else {
+            this.pendingPeoples.push(people);
+        }
+
+        this.form.markAsDirty();
+    }
+
+    private insertUnknownPeopleGroupAtBeginning(): void {
+        const unknown = this.pendingPeoples.find(p => p.identifier === UnknownPeopleGroup.PEID);
+        if (unknown) {
+            return;
+        }
+
+        const people = {} as PeopleAttributes;
+        people.identifier = UnknownPeopleGroup.PEID;
+        people.label = UnknownPeopleGroup.NmDisp;
+        people.placeOfOrigin = null;
+
+        const peopleField = this.template.getField('peoples');
+        peopleField.fields.forEach(field => {
+            people[field.id] = field.defaultValue || 0;
+        });
+
+        this.pendingPeoples.unshift(people);
+    }
+
+    public removePendingPeoples(people: PeopleAttributes): void {
+        const found = this.pendingPeoples.find(f => f.identifier === people.identifier && f.placeOfOrigin === people.placeOfOrigin);
+        const index = this.pendingPeoples.indexOf(found);
+        if (index > -1) {
+            this.pendingPeoples.splice(index, 1);
+        }
+
+        this.form.markAsDirty();
     }
 }