+++ src/templates/church-circles-czech/configuration-new.ts
@@ -135,9 +135,9 @@ export const ChurchCirclesCzechConfiguration: TemplateConfiguration = {
         },
         {
             id: 'attenders',
-            i18nRef: 'Common_NumberOfAttenders',
+            i18nRef: 'Common_NumberOfAttendees',
             defaultValue: 0,
-            type: ControlType.number,
+            type: ControlType.peopleSelector,
             canModify: true,
             iconRef: 'attendersIcon',
         },
@@ -145,7 +145,7 @@ export const ChurchCirclesCzechConfiguration: TemplateConfiguration = {
             id: 'believers',
             i18nRef: 'Common_NumberOfBelievers',
             defaultValue: 0,
-            type: ControlType.number,
+            // type: ControlType.number,
             canModify: true,
             iconRef: 'believersIcon',
         },
@@ -153,7 +153,7 @@ export const ChurchCirclesCzechConfiguration: TemplateConfiguration = {
             id: 'baptized',
             i18nRef: 'Common_NumberOfBaptized',
             defaultValue: 0,
-            type: ControlType.number,
+            // type: ControlType.number,
             canModify: true,
             iconRef: 'elementBaptismIcon'
         },
@@ -161,7 +161,7 @@ export const ChurchCirclesCzechConfiguration: TemplateConfiguration = {
             id: 'newlyBaptized',
             i18nRef: 'Common_NumberOfNewlyBaptized',
             defaultValue: 0,
-            type: ControlType.number,
+            // type: ControlType.number,
             canModify: true,
         },
         {
@@ -328,6 +328,53 @@ export const ChurchCirclesCzechConfiguration: TemplateConfiguration = {
             i18nRef: 'Common_Note',
             type: ControlType.textarea,
             canModify: true,
+        },
+        {
+            id: 'peoples',
+            canModify: false,
+            fields: [
+                {
+                    id: 'attenders',
+                    i18nRef: 'Common_NumberOfAttendees',
+                    defaultValue: 0,
+                    type: ControlType.number,
+                    canModify: true,
+                    iconRef: 'attendersIcon',
+                },
+                {
+                    id: 'believers',
+                    i18nRef: 'Common_NumberOfBelievers',
+                    defaultValue: 0,
+                    type: ControlType.number,
+                    canModify: true,
+                    iconRef: 'believersIcon',
+                    validation: {
+                        maxFieldRef: 'attenders',
+                    }
+                },
+                {
+                    id: 'baptized',
+                    i18nRef: 'Common_NumberOfBaptized',
+                    defaultValue: 0,
+                    type: ControlType.number,
+                    canModify: true,
+                    iconRef: 'elementBaptismIcon',
+                    validation: {
+                        maxFieldRef: 'believers',
+                    }
+                },
+                {
+                    id: 'newlyBaptized',
+                    i18nRef: 'Common_NumberOfNewlyBaptized',
+                    defaultValue: 0,
+                    type: ControlType.number,
+                    canModify: true,
+                    iconRef: 'elementBaptismIcon',
+                    validation: {
+                        maxFieldRef: 'baptized',
+                    }
+                }
+            ]
         }
     ],
     svgStates: [