+++ src/templates/countries.ts
@@ -3242,5 +3242,10 @@ export const COUNTRIES = [
         "region-code": "002",
         "sub-region-code": "202",
         "intermediate-region-code": "014"
+    },
+    {
+        "name": "Other",
+        "alpha-2": "Other",
+        "alpha-3": "Other"
     }
 ]